Transferred assets to Cloudinary database.
