--- dummy_data/InitTeam.ts
@@ -1,51 +1,0 @@
-export const team_members = [
-  {
-    name: "Julieta Vallejos",
-    position: "Executive Director, INIT",
-    linkedin: "https://www.linkedin.com/in/julieta-vallejos-pasco-61985790/",
-    image:
-      "https://uploads-ssl.webflow.com/62c343d1109fa5ddcec5736b/63f55271145df80e0ca054a6_1517436982527.jpg",
-  },
-  {
-    name: "Cesar Villa-Garcia",
-    position: " Product Manager, Microsoft",
-    linkedin: "https://www.linkedin.com/in/cesar-villa-garcia/",
-    image:
-      "https://uploads-ssl.webflow.com/62c343d1109fa5ddcec5736b/63da10ae762dccc95afcf9c4_1649133337418.jpeg",
-  },
-  {
-    name: "Yasmine Abdrabo",
-    position: " Site Reliability Engineer, Citi",
-    linkedin: "https://www.linkedin.com/in/yasmine-abdrabo/",
-    image:
-      "https://uploads-ssl.webflow.com/62c343d1109fa5ddcec5736b/63da10be2c8bfc706c24fe07_1563057292372.jpeg",
-  },
-  {
-    name: "Jose Maldonado",
-    position: " Software Engineer, Xbox",
-    linkedin: "https://www.linkedin.com/in/jmald045/",
-    image:
-      "https://uploads-ssl.webflow.com/62c343d1109fa5ddcec5736b/63da10ca7c3b01ad47edd4f9_1532962591994.jpeg",
-  },
-  {
-    name: "Sabina Cartacio",
-    position: " Product Manager, Fiddler",
-    linkedin: "https://www.linkedin.com/in/cartacios/",
-    image:
-      "https://uploads-ssl.webflow.com/62c343d1109fa5ddcec5736b/63da10b946c5801c70a7b2ee_1570342674567.jpeg",
-  },
-  {
-    name: "Christopher Rodriguez",
-    position: " Software Engineer, Microsoft",
-    linkedin: "https://www.linkedin.com/in/chrisrxyz/",
-    image:
-      "https://uploads-ssl.webflow.com/62c343d1109fa5ddcec5736b/63da10e4762dcc3c61fcfc37_1670015971027.jpeg",
-  },
-  {
-    name: "Andy Garcia",
-    position: " INIT FIU Chapter President",
-    linkedin: "https://www.linkedin.com/in/andygarcia-tech/",
-    image:
-      "https://uploads-ssl.webflow.com/62c343d1109fa5ddcec5736b/63da10eede0a10843640b44a_1632954731696.jpeg",
-  },
-];
