--- dummy_data/chapters.ts
@@ -1,26 +1,0 @@
-import FIU_GC from "/src/assets/images/FIU-GC.jpeg";
-import MDC_Wolfson from "/src/assets/images/MDC-Wolfson.png";
-import FMU_Campus from "/src/assets/images/FMU-Campus.jpeg";
-
-export const chapters = [
-  {
-    image: FIU_GC,
-    name: "Florida International University",
-    description: "Miami's public research university & leading HSI",
-  },
-  {
-    image: MDC_Wolfson,
-    name: "Miami Dade College",
-    description: "The most diverse educational institution in the nation",
-  },
-  {
-    image: FMU_Campus,
-    name: "Florida Memorial University",
-    description: "South Florida's HBCU & oldest academic center",
-  },
-  {
-    image: FIU_GC,
-    name: "Alumni / Professional",
-    description: "Alumni and professionals in the tech industry",
-  },
-];
