+++ src/Data/chapters.ts
@@ -1,0 +1,26 @@
+export const chapters = [
+  {
+    image:
+      "https://res.cloudinary.com/dh6y8bufo/image/upload/v1700204948/init-dashboard/about/chapters/FIU-GC.jpg",
+    name: "Florida International University",
+    description: "Miami's public research university & leading HSI",
+  },
+  {
+    image:
+      "https://res.cloudinary.com/dh6y8bufo/image/upload/v1700204952/init-dashboard/about/chapters/MDC-Wolfson.png",
+    name: "Miami Dade College",
+    description: "The most diverse educational institution in the nation",
+  },
+  {
+    image:
+      "https://res.cloudinary.com/dh6y8bufo/image/upload/v1700204950/init-dashboard/about/chapters/FMU-Campus.jpg",
+    name: "Florida Memorial University",
+    description: "South Florida's HBCU & oldest academic center",
+  },
+  {
+    image:
+      "https://res.cloudinary.com/dh6y8bufo/image/upload/v1700204948/init-dashboard/about/chapters/FIU-GC.jpg",
+    name: "Alumni / Professional",
+    description: "Alumni and professionals in the tech industry",
+  },
+];
